Solutions to 8th, 9th and 10th euler problems.

+++ common/primes.go
@@ -2,12 +2,12 @@ package common
 
 import "math"
 
-func IsPrime(n int) bool {
-	max := int(math.Sqrt(float64(n)))
+func IsPrime(n int64) bool {
+	max := int64(math.Sqrt(float64(n)))
 	if n < 2 {
 		return false
 	}
-	for i := 2; i <= max; i++ {
+	for i := int64(2); i <= max; i++ {
 		if n%i == 0 {
 			return false
 		}
